requirements.txt
requests==2.32.3
python-dateutil==2.9.0.post0
cryptography==44.0.2
urllib3==2.3.0
python-dotenv==1.0.1
websockets==14.1
datetime==5.5
